fix: Move components to components
