+++ src/routes/_app/route.tsx
@@ -3,7 +3,7 @@ import {
 	redirect,
 } from "@tanstack/react-router";
 import { getApiUsersCurrentOptions } from "lib/api/heyapi/@tanstack/react-query.gen";
-import AppLayout from "./-layout";
+import AppLayout from "./-components/-layout";
 
 export const Route = createFileRoute("/_app")({
 	beforeLoad: async ({ context, location }) => {
